resolve code review comments
