--- composeApp/src/commonMain/kotlin/com/qamar/petshelter/Greeting.kt
@@ -1,9 +1,0 @@
-package com.qamar.petshelter
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
