--- composeApp/src/commonMain/kotlin/com/qamar/petshelter/Platform.kt
@@ -1,9 +1,0 @@
-package com.qamar.petshelter
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
-
-expect fun getQamarFullName(): Platform
