+++ composeApp/src/commonMain/kotlin/com/qamar/petshelter/ui/home/ui/CategoriesSection.kt
@@ -12,6 +12,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.Text
@@ -46,7 +47,7 @@ fun CategoriesSection(
         Modifier.fillMaxWidth(),
         horizontalArrangement = Arrangement.spacedBy(16.dp, Alignment.CenterHorizontally)
     ) {
-        itemsIndexed(categoryList ?: listOf()) { index, item ->
+        items(categoryList ?: listOf()) { item ->
             item?.let {
                 CategoryItem(
                     item = item,
