--- composeApp/src/iosMain/kotlin/com/qamar/petshelter/Platform.ios.kt
@@ -1,12 +1,0 @@
-package com.qamar.petshelter
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
-actual fun getQamarFullName(): Platform {
-    TODO("Not yet implemented")
-}
